11/19分　エラーの処理を整頓

+++ app/Http/Controllers/DiaryController.php
@@ -57,6 +57,12 @@ class DiaryController extends Controller
         $diary = Diary::find($id);
         // dd($id);
 
+        //ログインユーザーが日記の投稿者かチェックする（他の人が内容を書き換えてしまうのを防ぐため）
+        if(Auth::user()->id !=$diary->user_id){
+            //投稿者とログインユーザーが違う場合
+            abort(403);
+        }
+
         //取得した日記の削除
         $diary->delete();
 
@@ -66,10 +72,20 @@ class DiaryController extends Controller
 
     //編集画面を表示する edit
     // index.blade.phpで使われてる
-    public function edit(int $id)
+    // public function edit(int $id)この書き方だとエラーが起こったときにプログラムコードのようなものが出てきて、一見壊れたようになってしまう。なので下のようにする。
+    //このときweb.phpの中身と同じにする。
+    public function edit(diary $diary)
     {
+
+        //ログインユーザーが日記の投稿者かチェックする（他の人が内容を書き換えてしまうのを防ぐため）
+        if(Auth::user()->id !=$diary->user_id){
+            //投稿者とログインユーザーが違う場合
+            abort(403);
+        }
+
+
         //受け取ったIDを元に日記を取得
-        $diary = Diary::find($id);
+        // $diary = Diary::find($id);
 
         //編集画面を返す。同時に取得した日記を渡す。
         return view('diaries.edit',[
